Migrate singly linked list
- Create required directories for implementation and tests
- Move files into proper directories
- Migrate singly_linked_list.cpp to header-only implementation
- Add unit tests for singly_linked_list.hpp
- Update README
- Update build script
