+++ C++/README.md
@@ -47,8 +47,8 @@ Implementation of well-known (and some rare) algorithms, in C++.
 
 ### Data structures
 * [Binary search tree](include/data_structures/binary_search_tree.cpp)
-* [Disjoint-set](include/data_structures/disjoint_set.cpp) :white_check_mark:
-* [Singly linked list](include/data_structures/singly_linked_list.cpp)
+* [Disjoint-set](include/data_structure/set/disjoint_set.hpp) :white_check_mark:
+* [Singly linked list](include/data_structure/linked_list/singly_linked_list.hpp) :white_check_mark:
 
 ## Compiling
 To compile the source files, run **`make all`**. Doing so will:
